Additional SHACL tests and tests for the SHACL tests themselves (#16)
I add the SHACL-based tests to test tree-specific sequences and whether certain tails are found within a particular "node", as well as a SPARQL-based test to check whether a node is listed as its own child. I also add a test script that deliberately constructs invalid trees to ensure that the SHACL tests catch them.