Adds purpose and overview to the front material
No scrolling or reading on should be needed to get a quick overview on what this tool does, hence putting that right up front in the README.

Also adds links to the CDAO and Phyloref ontologies. These will likely have to be changed once they're in their new maintained places.